Integrate changes back into eth-balance
