+++ development/states/account-detail-with-transaction-history.json
@@ -81,6 +81,20 @@
         "status": "confirmed",
         "containsDelegateCall": false,
         "hash": "0x957bbba51e2732a86c10c5e7e8a484a093795a06f2e2c38ad02da1b20aeca620"
+      },
+      {
+        "id": 1467921503489592,
+        "txParams": {
+          "from": "0x0dcd5d886577d5081b0c52e242ef29e70be3e7bc",
+          "to": "0x18a3462427bcc9133bb46e88bcbe39cd7ef0e761",
+          "value": "0x66c899104aa57038000",
+          "origin": "thelongestdomainnameintheworldandthensomeandthensomemoreandmore.com",
+          "metamaskId": 1467921503489592,
+          "metamaskNetworkId": "2"
+        },
+        "time": 1467921503489,
+        "status": "confirmed",
+        "containsDelegateCall": false
       }],
     "selectedAddress": "0x0dcd5d886577d5081b0c52e242ef29e70be3e7bc",
     "network": "2",
